Add logic package
