+++ src/main/java/seedu/duke/logic/command/GreetCommand.java
@@ -1,11 +1,11 @@
-package seedu.duke.command;
+package seedu.duke.logic.command;
 
 import seedu.duke.records.RecordList;
 import seedu.duke.records.biometrics.Biometrics;
-import seedu.duke.ui.Ui;
 import seedu.duke.records.exercise.ExerciseList;
 import seedu.duke.records.food.FoodList;
 import seedu.duke.storage.Storage;
+import seedu.duke.ui.Ui;
 
 public class GreetCommand extends Command {
