+++ src/main/java/seedu/duke/logic/command/LoadCommand.java
@@ -1,20 +1,21 @@
-package seedu.duke.command;
+package seedu.duke.logic.command;
 
 
-import seedu.duke.ui.Ui;
+import seedu.duke.exception.IllegalValueException;
 import seedu.duke.records.RecordList;
 import seedu.duke.records.biometrics.Biometrics;
-import seedu.duke.exception.IllegalValueException;
 import seedu.duke.records.exercise.ExerciseList;
 import seedu.duke.records.food.FoodList;
 import seedu.duke.storage.Storage;
+import seedu.duke.ui.Ui;
 
+import java.io.FileNotFoundException;
 import java.io.IOException;
 
 /**
- * Represents command for saving data into save file.
+ * Represents command for loading data from save file.
  */
-public class SaveCommand extends Command {
+public class LoadCommand extends Command {
 
     private Ui ui;
     private Storage storage;
@@ -22,8 +23,10 @@ public class SaveCommand extends Command {
     private ExerciseList exerciseList;
     private FoodList foodList;
 
+    private RecordList recordList;
 
-    public SaveCommand() {
+
+    public LoadCommand() {
     }
 
     @Override
@@ -34,15 +37,22 @@ public class SaveCommand extends Command {
         this.biometrics = biometrics;
         this.exerciseList = exerciseList;
         this.foodList = foodList;
+        this.recordList = recordList;
     }
 
     @Override
     public void execute() throws IllegalValueException {
         try {
-            storage.saveData(ui, biometrics, exerciseList, foodList);
-            ui.output("Saving data......");
-        } catch (IOException e) {
-            throw new IllegalValueException("unable to open save file");
+            storage.loadData(ui, biometrics, exerciseList, foodList, recordList);
+            ui.output("Remembering existing data......");
+        } catch (FileNotFoundException e) {
+            try {
+                storage.createDataFile();
+                ui.output("Data file created under data.txt");
+            } catch (IOException ex) {
+                throw new IllegalValueException("unable to create save file");
+            }
         }
     }
+
 }
