+++ src/main/java/seedu/duke/logic/command/SaveCommand.java
@@ -1,21 +1,20 @@
-package seedu.duke.command;
+package seedu.duke.logic.command;
 
 
-import seedu.duke.ui.Ui;
+import seedu.duke.exception.IllegalValueException;
 import seedu.duke.records.RecordList;
 import seedu.duke.records.biometrics.Biometrics;
-import seedu.duke.exception.IllegalValueException;
 import seedu.duke.records.exercise.ExerciseList;
 import seedu.duke.records.food.FoodList;
 import seedu.duke.storage.Storage;
+import seedu.duke.ui.Ui;
 
-import java.io.FileNotFoundException;
 import java.io.IOException;
 
 /**
- * Represents command for loading data from save file.
+ * Represents command for saving data into save file.
  */
-public class LoadCommand extends Command {
+public class SaveCommand extends Command {
 
     private Ui ui;
     private Storage storage;
@@ -23,10 +22,8 @@ public class LoadCommand extends Command {
     private ExerciseList exerciseList;
     private FoodList foodList;
 
-    private RecordList recordList;
 
-
-    public LoadCommand() {
+    public SaveCommand() {
     }
 
     @Override
@@ -37,22 +34,15 @@ public class LoadCommand extends Command {
         this.biometrics = biometrics;
         this.exerciseList = exerciseList;
         this.foodList = foodList;
-        this.recordList = recordList;
     }
 
     @Override
     public void execute() throws IllegalValueException {
         try {
-            storage.loadData(ui, biometrics, exerciseList, foodList, recordList);
-            ui.output("Remembering existing data......");
-        } catch (FileNotFoundException e) {
-            try {
-                storage.createDataFile();
-                ui.output("Data file created under data.txt");
-            } catch (IOException ex) {
-                throw new IllegalValueException("unable to create save file");
-            }
+            storage.saveData(ui, biometrics, exerciseList, foodList);
+            ui.output("Saving data......");
+        } catch (IOException e) {
+            throw new IllegalValueException("unable to open save file");
         }
     }
-
 }
