+++ src/test/java/seedu/duke/exercise/ExerciseListTest.java
@@ -1,15 +1,15 @@
 package seedu.duke.exercise;
 
 import org.junit.jupiter.api.Test;
-import seedu.duke.Parser;
-import seedu.duke.ui.Ui;
+import seedu.duke.exception.IllegalValueException;
+import seedu.duke.logic.Parser;
+import seedu.duke.logic.command.Command;
 import seedu.duke.records.RecordList;
 import seedu.duke.records.biometrics.Biometrics;
-import seedu.duke.command.Command;
-import seedu.duke.exception.IllegalValueException;
 import seedu.duke.records.exercise.ExerciseList;
 import seedu.duke.records.food.FoodList;
 import seedu.duke.storage.Storage;
+import seedu.duke.ui.Ui;
 
 import java.util.ArrayList;
